% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB.R
\name{calculateCBy}
\alias{calculateCBy}
\title{calculateCBy}
\usage{
calculateCBy(
  eef,
  CPy,
  approach = c("frequency", "weighing")[1],
  N = NULL,
  enddate = "date_end_monitoring_period",
  startdate = "date_start_monitoring_period"
)
}
\arguments{
\item{eef}{tibble}

\item{CPy}{tibble}

\item{approach}{Character. One of "frequency" or "weighing". Default: "frequency"}
}
\value{
tibble
}
\description{
calculateCBy
}
