% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP.R
\name{calculateCP}
\alias{calculateCP}
\title{calculateCP
Oorweeg of hierdie funksie korrek en nodig is !!!!!!!!!!! d}
\usage{
calculateCP(
  data,
  XPijk = "XPijk",
  XBijk = "XBijk",
  frMijk = "frMijk",
  frBPijk = "frBPijk",
  groupvar = "households",
  N = NULL,
  ...
)
}
\arguments{
\item{N}{}
}
\description{
calculateCP
Oorweeg of hierdie funksie korrek en nodig is !!!!!!!!!!! d
}
