% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP.R
\name{calculateCPy}
\alias{calculateCPy}
\title{calculateCPy}
\usage{
calculateCPy(
  CPy = NULL,
  dfFreq = NULL,
  datevars = c("year", "week", "weekday"),
  groupvar = "households",
  approach = c("frequency", "weighing")[1],
  indexvars = c("place", "year", "fuel"),
  N = NULL,
  enddate = "date_end_monitoring_period",
  startdate = "date_start_monitoring_period",
  frMijk = "frMijk",
  frMij = "frMij",
  XMijk = "XMijk",
  XMi = "XMi",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{CPy}{tibble}

\item{approach}{Character. One of "frequency" or "weighing". Default: "frequency"}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
tibble
}
\description{
calculateCPy
}
