% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BE.R
\name{calculateE}
\alias{calculateE}
\title{calculateE}
\usage{
calculateE(
  Cy,
  COEF = dfCOEF,
  fNRB = dffNRB,
  var = "CB",
  outcome = "BE",
  onlyOutcomeAndGroups = TRUE
)
}
\arguments{
\item{COEF}{Numeric. CO2 emission factor in g/kg.  Default value 1560}

\item{fNRB}{numeric. Default value 0.3}

\item{CBy}{Numeric. Result of calculateCBy()}
}
\value{
numeric
}
\description{
Generec emission calculation function
}
\examples{
calculateE(Cby)
}
