% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEEFi.R
\name{calculateEEFfromRR}
\alias{calculateEEFfromRR}
\title{calculatEEFfromRR}
\usage{
calculateEEFfromRR(
  data,
  indexvars = c("place", "year", "fuel"),
  XBij = "XBij_m2",
  XMij = "XMij_m2",
  frMij. = "frMij.",
  minfr = 5,
  rr = "rr",
  format = NULL,
  web3 = FALSE,
  ...
)
}
\arguments{
\item{data}{tibble or character}

\item{indexvars}{Character. One of c("place", "year", "fuel")}

\item{XBij}{Character. Average fuel use (in kg) per fire in baseline KT. Default "XBij_m3"}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}

\item{...}{}

\item{XMijk}{Character. Average fuel use (in kg) per brickstar fire in project KT. Default "XMij_m3"}
}
\description{
calculatEEFfromRR
}
