% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEF.R
\name{calculateEEFwithRR}
\alias{calculateEEFwithRR}
\title{calculatEEFwithRR}
\usage{
calculateEEFwithRR(
  data,
  indexvars = c("place", "year", "fuel"),
  groupvar = "households",
  XBijk = "XBijk",
  XMijk = "XMijk",
  ...
)
}
\arguments{
\item{data}{tibble}

\item{groupvar}{Character. Default "households"}

\item{XBijk}{Character. Default "XBijk"}

\item{XMijk}{Character. Default "XMijk"}

\item{...}{}
}
\description{
calculatEEFwithRR
}
