% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRR.R
\name{calculateFRR}
\alias{calculateFRR}
\title{fRR}
\usage{
calculateFRR(
  df = NULL,
  frBPij = "frBPij",
  frMij = "frMij",
  group_vars = c("place", "year", "fuel")
)
}
\arguments{
\item{df}{data.frame}

\item{frPij}{Character}

\item{fRMij}{Character}
}
\description{
fRR
}
