% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LE.R
\name{calculateLPy}
\alias{calculateLPy}
\title{calculateLPy}
\usage{
calculateLPy(
  CPy = NULL,
  CMsummary = CPfj_f,
  dfFreq = NULL,
  frr = NULL,
  group_vars = c("place", "year", "fuel"),
  N = NULL,
  qr = "household_qr_code",
  hhvar = "households",
  enddate = "date_end_monitoring_period",
  startdate = "date_start_monitoring_period",
  frMijk = "frMijk",
  frMij = "frMij",
  XBPijk = "XBPijk",
  XBPi = "XBPi"
)
}
\arguments{
\item{CPy}{tibble with summarise results of Project Kitchen test per household}

\item{CMsummary}{tibble with a summary of the Continued Monitoring data per household ¡¡¡
!!! consider creating internally from dfFreq}

\item{dfFreq}{tibble with the raw data of above}

\item{frr}{tibble with result from calculateFRR}

\item{group_vars}{Character. Grouping , typically c("place", "year", "fuel")}

\item{N}{Population size}

\item{qr}{Character. Variable in CPy with household ID or QR code}

\item{hhvar}{Character.}

\item{enddate}{Character. Variable in CMsummary with end of monitoring period}

\item{startdate}{Character. Variable in CMsummary with start of monitoring period}

\item{frMijk}{Character.}

\item{frMij}{Character.}

\item{XBPijk}{Character.}

\item{XBPi}{Character.}
}
\value{
tibble
}
\description{
calculateLPy
}
