% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{calculateRRj}
\alias{calculateRRj}
\title{calculateRRj}
\usage{
calculateRRj(
  data,
  groupvar = "households",
  frMijk = "frMijk",
  frBijk = "frBijk",
  frPijk = "frPijk",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
tibble with five column: <groupvar> meanfrBij meanfrMij meanfrPij    rr
}
\description{
Calculate replacement rate per household
}
