% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBi.R
\name{calculateXBi}
\alias{calculateXBi}
\title{calculateXBi}
\usage{
calculateXBi(
  data,
  XBijk = "XBijk",
  groupvar = "households",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{data}{tibble}

\item{XBijk}{Character. Variable with wood use from baseline technologies per day. Default "XBijk"}

\item{groupvar}{Character. Grouping variable. Default "households"}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
data.frame
}
\description{
Function to calculate baseline wood consumption from the baseline technologies per household
}
