% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XL.R
\name{calculateXl}
\alias{calculateXl}
\title{calculateXl}
\usage{
calculateXl(
  data,
  XPijk = "XPijk",
  XBijk = "XBijk",
  frMijk = "frMijk",
  frPijk = "frPijk",
  groupvar = "households",
  KPTValue = NULL,
  ...
)
}
\arguments{
\item{data}{tibble}

\item{XPijk}{Character. . Default "XPijk"}

\item{XBijk}{Character. . Default "XBijk"}

\item{frMijk}{Character. . Default "frMijk"}

\item{frPijk}{Character. . Default "frPijk"}

\item{groupvar}{Character. Grouping variable. Default "households"}

\item{KPTValue}{Numeric or NULL. If NULL it is calculated from the same dataset}

\item{...}{Arguments passed to simulateXBi}
}
\value{
data.frame
}
\description{
Function to calculate leakage wood use, i.e. wood use from baseline technologies in the
project scenario. This is done by applying the fuel per fire derived from the baseline
kitchen test and the frequency of fires using the baseline technologies in the project scenario.
}
