% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMB.R
\name{calculateXMB}
\alias{calculateXMB}
\title{calculateXMB}
\usage{
calculateXMB(
  data,
  groupvar = "households",
  XBijk = "XBijk",
  XMijk = "XMijk",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{groupvar}{Character. Grouping variable. Default: "households"}

\item{XBijk}{Character. Default: "XBijk"}

\item{XMijk}{Character. Default: "XMijk"}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
tibble with one row and two columns namely "xBi" and "xMi"
}
\description{
Calculate xBi xMi
}
