% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfs2R.R
\name{ipfsDF}
\alias{ipfsDF}
\title{ipfsDF}
\usage{
ipfsDF(df = dfIPFS, nn = "ER")
}
\arguments{
\item{nn}{Character.  Name of object to recover}

\item{dfIPFS}{tibble. Typically resulting from ls2ipfs()}
}
\value{
R Object
}
\description{
Convenience function to use dfIPFS to recover an R object from IPFS
}
\examples{
ER = ipfsDF("ER")  # same as  ipfsDF(df = dfIPFS, nn = "ER")
}
