% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfs2R.R
\name{ls2ipfs}
\alias{ls2ipfs}
\title{ls2ipfs
Add all noin-function objects in the Global Environment to IPFS}
\usage{
ls2ipfs()
}
\value{
tibble
}
\description{
ls2ipfs
Add all noin-function objects in the Global Environment to IPFS
}
\examples{
ls2ipfs()
}
