% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dfCBi.R
\name{make_dfCBi}
\alias{make_dfCBi}
\title{make_dfCBi}
\usage{
make_dfCBi(
  dfEEF = dfRR_EEFij,
  dfCP = dfCPi,
  frMij. = "frMij.",
  indexvars = c("place", "year", "fuel"),
  eef = "eef",
  CPi = "CPi",
  minfr = 5,
  N = 1000,
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{indexvars}{list with characters names. List of variables that uniquely identify subpopulation}

\item{N}{number of households that was monitored during continuous monitoring phase}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}

\item{dfresP}{tibble. Df with the results from Project Kitchen Test}

\item{dfConMonFreq}{tibble. Df with the results from Continuous Monitoring phase}

\item{groupvar}{character. household ID.}

\item{date.}{Date of continuous monitoring per household.}

\item{approach}{Character. One of "frequency" or "weighing". Default: "frequency"}
}
\value{
tibble
}
\description{
make_dfCBi
}
