% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dfCPi.R
\name{make_dfCPi}
\alias{make_dfCPi}
\title{make_dfCPi}
\usage{
make_dfCPi(
  dfConMon = dfConMon,
  KTresults = dfresBP,
  indexvars = c("place", "year", "fuel"),
  startdate = "date_start_monitoring_period",
  enddate = "date_end_monitoring_period",
  N = 19,
  ndays_cm = "ndays_cm",
  nfire_cm = "nfire_cm",
  XMij = "XMij_m2",
  XBij = "XBij_m2",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{indexvars}{list with characters names. List of variables that uniquely identify subpopulation}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
tibble
}
\description{
make_dfCPi
}
