% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dfFrRij.R
\name{make_dfFrRij}
\alias{make_dfFrRij}
\title{make_dfFrRij}
\usage{
make_dfFrRij(
  df = dfresP,
  indexvars = c("place", "year", "fuel"),
  frMij. = "frMij.",
  frBPij. = "frBPij.",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
tibble with five column: <groupvar> meanfrBij meanfrMij meanfrPij    rr
}
\description{
Calculates the average frequency ratio of baseline fires to project measures fires in the project phase of the KT
}
