% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dfRR_EFFij.R
\name{make_dfRRij_EEFij}
\alias{make_dfRRij_EEFij}
\title{make_dfRRij_EEFij}
\usage{
make_dfRRij_EEFij(
  data,
  frBij. = "frBij.",
  frMij. = "frMij.",
  frBPij. = "frBPij.",
  XBij = "XBij_m2",
  XMij = "XMij_m2",
  ndays_BKT = "ndays_BKT",
  ndays_PKT = "ndays_PKT",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
tibble with five column: <groupvar> meanfrBij meanfrMij meanfrPij    rr
}
\description{
Calculate replacement rate per household
}
