% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDateRef.R
\name{makeDateRef}
\alias{makeDateRef}
\title{makeDateRef}
\usage{
makeDateRef(startdate, enddate, format = NULL, web3 = FALSE)
}
\arguments{
\item{startdate}{Date}

\item{enddate}{Date}

\item{format}{Character. Output format for API}

\item{web3}{Return a web3storage address or not. Default FALSE}
}
\value{
tibble
}
\description{
Create a reference table with every day of a certain period
}
\examples{
makeDateRef(startdate = as.Date("2022-01-01"), enddate = as.Date("2022-12-31"))
}
