% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedfConMonSD.R
\name{makedfConMonSD}
\alias{makedfConMonSD}
\title{makedfConMonSD}
\usage{
makedfConMonSD(
  df = NULL,
  dfHouse = NULL,
  ignorevalue = 0,
  minobs = 24,
  startdatetime = ymd_hms("2023-06-22 00:00:00"),
  returnFull = FALSE,
  valuevar = "stove_temp",
  sensor_id = "sensor_id",
  datevar = "date",
  placevar = "place",
  fileidvar = "file_id",
  sensor_locationvar = "sensor_location",
  id_var = "qr_code_household",
  refvar = "box_temp",
  fuelvar = "fuel"
)
}
\arguments{
\item{df}{tibble. Output of readSD}

\item{dfHouse}{tibble with house - instrument pairings}

\item{ignorevalue}{Numeric. Value to ignore indicating sensor not working. Default is 0}

\item{minobs}{Numeric. Minimum number of observations to consider a day. Default is 24}

\item{startdatetime}{Daretime. Start date for the data. Default is 2023-06-22 00:00:00}

\item{returnFull}{Logical. If TRUE, returns the full dataframe with all columns. If false, returns only dfConMon format. Default is FALSE}

\item{valuevar}{Character. Name of variable with temperature. Default is "stove_temp"}

\item{sensor_id}{Character. Name of variable with sensor ID. Default is "sensor_id"}

\item{datevar}{Character. Name of variable with date variable}

\item{placevar}{Character. Name of variable with place. Default is "place"}

\item{fileidvar}{Character. Name of variable with file ID}

\item{sensor_locationvar}{Character. Name of variable with sensor location. Default is "sensor_location"}

\item{id_var}{Character. Name of variable with household ID}

\item{refvar}{Character. Name of variable with reference temperature. Default is "box_temp"}

\item{fuelvar}{Character. Name of variable with fuel. Default is "fuel"}
}
\value{
tibble
}
\description{
Get output of readSD into a dataframe called dfNest with colums: sensor_location, qr_code_household, data
data has columns:  date <dttm>, temp <dbl>,  ibutton_id <chr> file_id <chr>
}
\examples{
dft <- decodeSDcards(path = "Data/SDcards")
dfTemp <- dft \%>\% select(file_id, sensor_id, data12) \%>\% unnest(data12)
makedfConMonSD(df = dfTemp, dfHouse = dfSDhouse \%>\% filter(!is.na(place)),datevar = "date", fileidvar = "file_id", sensor_location = "sensor_location", id_var = "qr_code_household", valuevar = "stove_temp")

}
