% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTablesforER.R
\name{makeTablesforER}
\alias{makeTablesforER}
\title{makeTablesforER}
\usage{
makeTablesforER(
  dfConMon = NULL,
  dfresBP = NULL,
  startdate = lubridate::ymd("2023-06-22"),
  enddate = lubridate::ymd("2023-08-08"),
  dfCOEF = tibble(fuel = c("wood", "charcoal"), COEF = c(units::as_units(1560, "g/kg"),
    units::as_units(2860, "g/kg"))),
  dffNRB = tibble(place = "Lwandlamuni", year = c(2023, 2024), fNRB = c(0.3, 0.3)),
  path = "C:/Users/User/Documents/NOVA/HGICPcalculator/"
)
}
\arguments{
\item{dfConMon}{df. Data frame with continuous monitoring data}

\item{dfresBP}{df. Data frame with combined results for the Baseline Kitchen Test and the Project Kitchen Test}

\item{startdate}{date. Start date for continuous monitoring}

\item{enddate}{date. End date for continuous monitoring}
}
\description{
Calculates the tables that are included in the ER report
}
