% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCPij.R
\name{plotCPij}
\alias{plotCPij}
\title{plotCPij}
\usage{
plotCPij(
  CPij,
  place = "place",
  year = "year",
  fuel = "fuel",
  assignment = "assignment",
  household_qr_code = "household_qr_code",
  kgvars = c("XMij", "XBPijk"),
  monthvars = c("kg_p_month_m2k"),
  freqvars = c("frBPij", "frMij")
)
}
\arguments{
\item{CPij}{tibble}

\item{place}{Character. Name of variable indicating the place}

\item{year}{Character. Name of variable indicating the year}

\item{fuel}{Character. Name of variable indicating the fuel type}

\item{assignment}{Character. Name of variable indicating the Kitchen test phase}

\item{household_qr_code}{Character. Name of variable indicating the household ID}

\item{kgvars}{Character. Name of variable measured in kg per day (XMij", "XBPijk)}

\item{monthvars}{Character. Name of variable indicating the monthly wood use}

\item{freqvars}{Character. Name of variable measured in frequency per day}
}
\value{
plotly
}
\description{
plotCPij
}
