% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web3S2R.R
\name{R2web3S}
\alias{R2web3S}
\title{R2web3S}
\usage{
R2web3S(x, AT = Sys.getenv("IPFS_STORAGE_API_KEY"), fnmW = NULL)
}
\arguments{
\item{x}{R object}

\item{AT}{Character. The 255 character IPFS storage key. By default it Gets"IPFS_STORAGE_API_KEY" from Sytem Environment}

\item{fnmW}{A file. if NULL (the default), a tempfile is created}
}
\value{
Character. The IPFS URI where the result is writen
}
\description{
R2web3S
}
