% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFromPKT.R
\name{resultsFromPKT}
\alias{resultsFromPKT}
\title{Title}
\usage{
resultsFromPKT(
  projectKT,
  place = "place",
  year = "year",
  households = "households",
  frPijk = "frPijk",
  frMijk = "frMijk",
  frBPijk = "frBPijk",
  XMijk = "XMijk",
  XBPijk = "XBPijk",
  XPijk = "XPijk",
  fuel = "fuel",
  fuelval = "wood",
  assignment = "assignment",
  assignmentval = "projectKT",
  household_qr_code = "household_qr_code",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{projectKT}{tibble or character. If character assumes it is a web3storage address of a tibble}

\item{place}{Character. Column name in projectKT with place}

\item{year}{Character. Column name in projectKT with year}

\item{households}{Character. Column name in projectKT with initial household ID}

\item{frPijk}{Character. Column name in projectKT with frequency of wood use in project scenario}

\item{frMijk}{Character. Column name in projectKT with frequency of wood use in project devices in the project scenario}

\item{frBPijk}{Character. Column name in projectKT with frequency of wood use in baseline devices project scenario}

\item{XMijk}{Character. Column name in projectKT with mean wood use in project devices in the project scenario}

\item{XBPijk}{Character. Column name in projectKT with mean wood use in baseline devices in the project scenario}

\item{XPijk}{Character. Column name in projectKT with mean wood use in ???? all ???? devices in the project scenario}

\item{fuel}{Character. Column name to be created for fuel type}

\item{fuelval}{Character. Value to set fuel to. Default "wood"}

\item{assignment}{Character. Column name to be created for assignment}

\item{assignmentval}{Character. Value to set fuel to. Default "projectKT"}

\item{household_qr_code}{Character. Name to use for the ID column. Default household_qr_code}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Logical. Return web3storage address or not}
}
\description{
Title
}
