% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializer_switch.R
\name{serializer_switch}
\alias{serializer_switch}
\title{serializer_switch}
\usage{
serializer_switch()
}
\value{
function
}
\description{
serializer_switch
}
\examples{
register_serializer("switch", serializer_switch)
}
