% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB.R
\name{simulateCB}
\alias{simulateCB}
\title{Title}
\usage{
simulateCB(
  data,
  XPijk = "XPijk",
  XBijk = "XBijk",
  frMijk = "frMijk",
  frBPijk = "frBPijk",
  frBijk = "frBijk",
  frPijk = "frPijk",
  groupvar = "households",
  stratvar = "Z",
  NDPi = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
