% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP.R
\name{simulateCP}
\alias{simulateCP}
\title{Title}
\usage{
simulateCP(
  data,
  XPijk = "XPijk",
  XBijk = "XBijk",
  frMijk = "frMijk",
  frBPijk = "frBPijk",
  groupvar = "households",
  N = NDP,
  ...
)
}
\arguments{
\item{N}{}
}
\description{
Title
}
