% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER.R
\name{simulateER}
\alias{simulateER}
\title{simulateER}
\usage{
simulateER(data, format = NULL, web3 = FALSE, ...)
}
\arguments{
\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
data.frame
}
\description{
simulateER
}
