% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrM.R
\name{simulateFrM}
\alias{simulateFrM}
\title{simulateFrM}
\usage{
simulateFrM(data, selection_statement = "Z_kt == 1 & Z_pw == 1", ...)
}
\arguments{
\item{data}{tibble}

\item{...}{passed to calculateFrM}
}
\value{
data.frame
}
\description{
Simulation function for calculateFrM. Filters for kitchen test group and project scenario week
}
\examples{
est_Frm <- declare_estimators(handler = label_estimator(simulateFrM), label = "frM", inquiry = "aveFrM")
}
