% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{simulateRRj}
\alias{simulateRRj}
\title{simulateRRj}
\usage{
simulateRRj(data, selection_statement = "Z_kt == 1 &  Z_pw == 1", ...)
}
\arguments{
\item{selection_statement}{Character. filter exporession: Default :"Z_kt == 1 &  Z_pw == 1"}

\item{...}{}
}
\description{
simulateRRj
}
