% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBi.R
\name{simulateXBi}
\alias{simulateXBi}
\title{simulateXBi}
\usage{
simulateXBi(data, selection_statement = "Z_kt == 1 &  Z_bw == 1", ...)
}
\arguments{
\item{...}{}
}
\description{
Function to use in simulation instead of calculateXBi.
Filters data for kitchen test group (Z_kt == 1) and  baseline week (Z_bw == 1)
}
\examples{
est_XBi <- declare_estimator(handler = label_estimator(simulateXBi), label = "XBi", inquiry = "XBi")
}
