% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XL.R
\name{simulateXl}
\alias{simulateXl}
\title{simulateXl}
\usage{
simulateXl(
  data,
  selection_statement1 = "Z_kt == 1 & Z_bw == 1",
  selection_statement2 = "Z_cm == 1 & Z_cmw == 1",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Function to simulate Leakage use for instead of calculateXl with simulated data.
Isolates baseline kitchen test households and period for the calculation of baseline wood
consumption per fire and then isolates the continuous monitoring households in continuous
monitoring phase for calculation fo the frequency of fires in the project scenario
}
