% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMB.R
\name{simulateXMB}
\alias{simulateXMB}
\title{simulateXMB}
\usage{
simulateXMB(data, selection_statement = "Z_kt == 1 &  Z_pw == 1", ...)
}
\arguments{
\item{...}{}

\item{selection_statement.}{Character representation of filtering expression:
Default:  "Z_kt == 1 &  Z_pw == 1"}
}
\value{
tibble
}
\description{
simulateXMB
}
