% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMi.R
\name{simulateXMi}
\alias{simulateXMi}
\title{simulateXMi}
\usage{
simulateXMi(data, selection_statement = "Z_kt == 1 &  Z_pw == 1", ...)
}
\arguments{
\item{data}{tibble}

\item{selection_statement}{Character string of expression to be evaluated to select data.
Dafault "Z_kt == 1 &  Z_pw == 1"}

\item{...}{passed to calculateXMi}
}
\value{
data.frame
}
\description{
Function for use with simulated data selecting the kitchen test group (Z_kt == 1)
and the project scenario test phase (Z_pw == 1)
}
\examples{
est_XMi <- declare_estimator(handler = label_estimator(simulateXMi), label = "XMi", inquiry = "XMi")
}
