% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaiseConMon.R
\name{summariseConMon}
\alias{summariseConMon}
\title{summariseConMon}
\usage{
summariseConMon(
  dfFreqRes_prep = NULL,
  household_qr_code = "household_qr_code",
  date = "date",
  frMijk = "frMijk",
  place = "place",
  year = "year",
  fuel = "fuel",
  assignment = "assignment",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{dfFreqRes_prep}{tibble or character. Assumes a web3Storage address if a character}

\item{household_qr_code}{Character. Column in dfFreqRes_prep with household ID}

\item{date}{Character. Column in dfFreqRes_prep with date}

\item{frMijk}{Character. Column in dfFreqRes_prep with fires per day from project measures}

\item{place}{Character. Column in dfFreqRes_prep with place}

\item{year}{Character. Column in dfFreqRes_prep with year}

\item{fuel}{Character. Column in dfFreqRes_prep with fuel}

\item{assignment}{Character. Column in dfFreqRes_prep with assignment (i.e. monitoring phase)}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}
}
\description{
Summarise a tibble with daily fire counts for project measures to a total and average count for the monitoring period
}
\examples{
summariseConMon("https://bafkreiasme5plzoyimo3k2r33tpbtjndsf7ykch7ihkctx5em23fffskza.ipfs.w3s.link/")
}
