% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchify.R
\name{switchify}
\alias{switchify}
\title{switchify}
\usage{
switchify(out, format = NULL)
}
\arguments{
\item{out}{tibble}

\item{format}{Character of NULL}
}
\value{
tibble
}
\description{
Use the serialiser_switch function and the "switch" serializer to switch between json. csv and html
serialisations when using function as a API
}
\examples{
result \%>\% switchify()
}
