% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3lify.R
\name{web3lify}
\alias{web3lify}
\title{Title}
\usage{
web3lify(x, web3 = FALSE)
}
\arguments{
\item{x}{object}

\item{web3}{Logical. If FALSE (default) returns x, else
returns a web3storage address resulting from R2web3S(x)}
}
\description{
Title
}
\examples{
iris \%>\% web3lify(web3 = TRUE)
}
