% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web3S2R.R
\name{web3S2R}
\alias{web3S2R}
\title{web3S2R
Resolve IPFS https url and retrive the Rds file}
\usage{
web3S2R(ipfsURI, fnmR = NULL)
}
\arguments{
\item{fnmR}{Character. If NULL generates a tempfile}
}
\value{
R object
}
\description{
web3S2R
Resolve IPFS https url and retrive the Rds file
}
\examples{
web3S2R("https://bafkreigbspthbmhuza3ypk4vzdcvzclmt4kqusdybepxl54kjocjgcsdla.ipfs.w3s.link/")
}
