% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web3Sub.R
\name{web3Sub}
\alias{web3Sub}
\title{web3Sub}
\usage{
web3Sub(x)
}
\arguments{
\item{x}{}
}
\description{
web3Sub
}
\examples{
z <- R2web3S(iris)
web3Sub(x = "https://bafkreiabt5id676i65g2f25quzub2ib2ecz6xn6xzyalp7opmhvg7ybxum.ipfs.w3s.link/")
}
