% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB.R
\name{calculateCBy}
\alias{calculateCBy}
\title{calculateCBy}
\usage{
calculateCBy(
  eef,
  CPy,
  approach = c("frequency", "weighing")[1],
  N = NULL,
  enddate = "date_end_monitoring_period",
  startdate = "date_start_monitoring_period",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{eef}{tibble}

\item{CPy}{tibble}

\item{approach}{Character. One of "frequency" or "weighing". Default: "frequency"}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
tibble
}
\description{
calculateCBy
}
