% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateE.R
\name{calculateE}
\alias{calculateE}
\title{calculateE}
\usage{
calculateE(
  Cy,
  COEF = dfCOEF,
  fNRB = dffNRB,
  var = "CBi",
  outcome = "BE",
  onlyOutcomeAndGroups = TRUE,
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{COEF}{Numeric. CO2 emission factor in g/kg.  Default value 1560}

\item{fNRB}{numeric. Default value 0.3}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}

\item{CBy}{Numeric. Result of calculateCBy()}
}
\value{
numeric
}
\description{
Generec emission calculation function
}
\examples{
calculateE(Cby)
}
