% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEF.R
\name{calculateEEFwithRR}
\alias{calculateEEFwithRR}
\title{calculatEEFwithRR}
\usage{
calculateEEFwithRR(
  data,
  indexvars = c("place", "year", "fuel"),
  groupvar = "households",
  XBijk = "XBijk",
  XMijk = "XMijk",
  format = NULL,
  web3 = FALSE,
  ...
)
}
\arguments{
\item{data}{tibble or character}

\item{indexvars}{Character. One of c("place", "year", "fuel")}

\item{groupvar}{Character. Default "households"}

\item{XBijk}{Character. Default "XBijk"}

\item{XMijk}{Character. Default "XMijk"}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}

\item{...}{}
}
\description{
calculatEEFwithRR
}
