% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER.R
\name{calculateER}
\alias{calculateER}
\title{calculateER}
\usage{
calculateER(BE, PE, LE)
}
\arguments{
\item{BE}{tibble from calculateE(  ..., outcome = "BE")}

\item{PE}{tibble from calculateE(  ..., outcome = "PE")}

\item{LE}{tibble from calculateE(  ..., outcome = "LE")}

\item{onlyOutcomeAndGroups}{Logical}
}
\value{
data.frame
}
\description{
calculateER
}
