% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrM.R
\name{calculateFrM}
\alias{calculateFrM}
\title{calculateFrM}
\usage{
calculateFrM(
  data,
  frMijk = "frMijk",
  groupvar = "households",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{data}{tibble}

\item{frMijk}{Character. Variable with frequency of fire from project technology
by household group, household and day. Default "frMijk"}

\item{groupvar}{Character. Variable for grouping. Default "households"}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\value{
data.frame
}
\description{
calculate the mean frequency of use of the project measures by households
}
