% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRR.R
\name{calculateFRR}
\alias{calculateFRR}
\title{fRR}
\usage{
calculateFRR(
  df = NULL,
  frBPij = "frBPij",
  frMij = "frMij",
  group_vars = c("place", "year", "fuel"),
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{df}{data.frame}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}

\item{frPij}{Character}

\item{fRMij}{Character}
}
\description{
fRR
}
