% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{calculateRRj}
\alias{calculateRRj}
\title{calculateRRj}
\usage{
calculateRRj(
  data,
  groupvar = "households",
  frMijk = "frMijk",
  frBijk = "frBijk",
  frPijk = "frPijk"
)
}
\arguments{
\item{frPijk}{}
}
\value{
tibble with five column: <groupvar> meanfrBij meanfrMij meanfrPij    rr
}
\description{
Calculate replacement rate per household
}
