% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMi.R
\name{calculateXMi}
\alias{calculateXMi}
\title{calculateXMi}
\usage{
calculateXMi(data, XMijk = "XMijk", groupvar = "households")
}
\arguments{
\item{data}{tibble}

\item{XMijk}{Character. Variable with daily wood use per household. Default "XMijk"}

\item{groupvar}{Character. Grouping variable. Default "households"}
}
\value{
data.frame
}
\description{
Function to calculate the mean wood use by the project measures by household
}
