% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfs2R.R
\name{ls2ipfs}
\alias{ls2ipfs}
\title{ls2ipfs
Add all non-function objects in the Global Environment to IPFS}
\usage{
ls2ipfs(storage = c("IPFS", "WEB3STORAGE")[2])
}
\value{
tibble
}
\description{
ls2ipfs
Add all non-function objects in the Global Environment to IPFS
}
\examples{
ls2ipfs()
}
