% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dfCPi.R
\name{make_dfCPi}
\alias{make_dfCPi}
\title{make_dfCPi}
\usage{
make_dfCPi(
  dfConMonFreq = dfFreqRes_prep,
  KTresults = dfresBP,
  indexvars = c("place", "year", "fuel"),
  groupvar = "household_qr_code",
  frMijk = "frMijk",
  XMij = "XMij_m2",
  XBij = "XBij_m2",
  date = "date",
  N = 1000,
  enddate = "date_end_monitoring_period",
  startdate = "date_start_monitoring_period",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{dfConMonFreq}{tibble. Df with the results from Continuous Monitoring phase}

\item{indexvars}{list with characters names. List of variables that uniquely identify subpopulation}

\item{groupvar}{character. household ID.}

\item{N}{number of households that was monitored during continuous monitoring phase}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}

\item{dfresP}{tibble. Df with the results from Project Kitchen Test}

\item{date.}{Date of continuous monitoring per household.}

\item{approach}{Character. One of "frequency" or "weighing". Default: "frequency"}
}
\value{
tibble
}
\description{
make_dfCPi
}
