% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDateRef.R
\name{makeDateRef}
\alias{makeDateRef}
\title{makeDateRef}
\usage{
makeDateRef(startdate, enddate)
}
\arguments{
\item{startdate}{Date}

\item{enddate}{Date}
}
\value{
tibble
}
\description{
Create a reference table with every day of a certain period
}
\examples{
makeDateRef(startdate = as.Date("2022-01-01"), enddate = as.Date("2022-12-31"))
}
