% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDateRef.R
\name{makeDateRef}
\alias{makeDateRef}
\title{makeDateRef}
\usage{
makeDateRef(
  ERstartdate = as.Date("2023-01-01"),
  ERenddate = as.Date("2023-05-31")
)
}
\arguments{
\item{ERstartdate}{Date}

\item{ERenddate}{Date}
}
\value{
tibble
}
\description{
makeDateRef
}
