% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFromBKT.R
\name{resultsFromBKT}
\alias{resultsFromBKT}
\title{resultsFromBKT}
\usage{
resultsFromBKT(
  baselineKT = NULL,
  place = "place",
  year = "year",
  households = "households",
  XBijk = "XBijk",
  frBijk = "frBijk",
  fuel = "fuel",
  fuelval = "wood",
  assignment = "assignment",
  assignmentval = "baselineKT",
  format = NULL,
  web3 = FALSE
)
}
\arguments{
\item{baselineKT}{tibble or character. If character assumes it is a web3storage address of a tibble}

\item{place}{Character}

\item{year}{Character}

\item{households}{Character}

\item{XBijk}{Character}

\item{frBijk}{Character}

\item{fuelval}{Character}

\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}

\item{assval}{Character}

\item{household_qr_code}{Character}
}
\value{
tibble
}
\description{
resultsFromBKT
}
\examples{
x <- "https://bafkreib7u7afwxo4gwyn46nlpuxuh7zzacqcgsy5p6nop75xskmrqlnbpe.ipfs.w3s.link/"
resultsFromBKT(x, format = "htmlTable")
}
