% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB.R
\name{simulateCB}
\alias{simulateCB}
\title{simulateCB}
\usage{
simulateCB(
  data,
  XPijk = "XPijk",
  XBijk = "XBijk",
  frMijk = "frMijk",
  frBPijk = "frBPijk",
  frBijk = "frBijk",
  frPijk = "frPijk",
  groupvar = "households",
  stratvar = "Z",
  NDPi = NULL,
  ...
)
}
\arguments{
\item{data}{tibble}

\item{XPijk}{Character. Default "XPijk"}

\item{XBijk}{Character. Default "XBijk"}

\item{frMijk}{Character. Default  "frMijk"}

\item{frBPijk}{Character. Default "frBPijk"}

\item{frBijk}{Character. Default "frBijk"}

\item{frPijk}{Character. Default "frPijk"}

\item{groupvar}{Character. Default "groupvar"}

\item{stratvar}{Character. Default "stratvar"}

\item{...}{passed to simulateCP}
}
\value{
tibble
}
\description{
simulateCB
}
