% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP.R
\name{simulateCP}
\alias{simulateCP}
\title{simulateCP}
\usage{
simulateCP(
  data,
  XPijk = "XPijk",
  XBijk = "XBijk",
  frMijk = "frMijk",
  frBPijk = "frBPijk",
  groupvar = "households",
  N = NDP,
  format = NULL,
  web3 = FALSE,
  ...
)
}
\arguments{
\item{format}{Character. Output format if called as API. one of "json", "csv", "rds", "htmlTable"}

\item{web3}{Logical. Return web3storage address or not}
}
\description{
simulateCP
}
